requirements.txt
numpy==1.24.3
opencv-python==4.8.0.76
pympi-ling==1.70.2
pandas==2.1.0
gitpython==3.1.35
pyyaml==6.0.1
toml==0.10.2
scikit-image==0.21.0
h5py==3.9.0
matplotlib
myst_parser
